@@ -1,6 +1,5 @@
-'use strict';
-const escapeStringRegexp = require('escape-string-regexp');
-const mapping = require('./lib/mappings');
+import escapeStringRegexp from 'escape-string-regexp';
+import mapping from './lib/mappings.js';
 
 const hasFlags = (regexFlags, replaceFlags) => {
 	if (!replaceFlags) {
@@ -11,9 +10,7 @@ const hasFlags = (regexFlags, replaceFlags) => {
 	return replaceFlags.split('').every(flag => regexFlags.includes(flag));
 };
 
-module.exports = (regexp, flags) => {
-	flags = flags || '';
-
+export default function cleanRegexp(regexp, flags = '') {
 	if (typeof regexp !== 'string') {
 		throw new TypeError(`Expected regexp to be of type \`string\`, got \`${typeof regexp}\``);
 	}
@@ -22,14 +19,11 @@ module.exports = (regexp, flags) => {
 		throw new TypeError(`Expected flags to be of type \`string\`, got \`${typeof flags}\``);
 	}
 
-	for (const replace of mapping) {
-		const key = replace[0];
-		const replacement = replace[1];
-
+	for (const [key, replacement] of mapping) {
 		if (hasFlags(flags, replacement.flags)) {
 			regexp = regexp.replace(new RegExp(escapeStringRegexp(key), 'g'), replacement.value);
 		}
 	}
 
 	return regexp;
-};
+}
