@@ -1,6 +1,4 @@
-'use strict';
-
-module.exports = new Map([
+export default new Map([
 	['[0-9]', {value: '\\d'}],
 	['[^0-9]', {value: '\\D'}],
 
@@ -139,5 +137,5 @@ module.exports = new Map([
 	['[^\\da-z_]', {value: '\\W', flags: 'i'}],
 	['[^\\d_a-z]', {value: '\\W', flags: 'i'}],
 	['[^_a-z\\d]', {value: '\\W', flags: 'i'}],
-	['[^_\\da-z]', {value: '\\W', flags: 'i'}]
+	['[^_\\da-z]', {value: '\\W', flags: 'i'}],
 ]);
